Improve startup time
