+++ styles/dnote-atom.less
@@ -4,5 +4,10 @@
 // for a full listing of what's available.
 @import "ui-variables";
 
-.dnote-atom {
+.dnote-tile {
+  display: inline-block;
+
+  &:hover {
+    background-color: #dbdbdc;
+  }
 }
